refactor: Move tests into modules
